Revert "PowerMax driver - rebranding VMAX to PowerMax"
The patch was not backward compatible. We should properly
deprecate the old parameters to not break our users.

This reverts commit 0c831bbed88fa3cc3177b76b94738fc447b06a2f.
