--- releasenotes/notes/powermax-rebrand-driver-d428e372280c44e5.yaml
@@ -1,5 +1,0 @@
----
-upgrade:
-  - |
-    PowerMax driver - 'PowerMax' is to replace 'VMAX' in release notes,
-    online documentation, directory structure and driver name.
